requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
jinja2
openai
python-dotenv
pytest
httpx
alembic
cryptography
python-multipart
requests